+++ apps/insights/app/blog/cloudflare.tsx
@@ -1,5 +1,5 @@
 import { AreaChart } from '@/components/charts'
-import { CompactMetric } from '@/components/ui/compact-metric'
+import { CompactMetric } from '@/components/ui/CompactMetric'
 import type { CloudflareAnalyticsByDate } from '@duyet/interfaces'
 import { request } from 'graphql-request'
 import { Activity, Eye, Globe, Users } from 'lucide-react'
