+++ apps/insights/app/github/components/index.ts
@@ -1,0 +1,3 @@
+// Re-export all GitHub components
+export { CommitMetrics } from './CommitMetrics'
+export { CommitChart } from './CommitChart'
