+++ apps/insights/app/github/language-stats.tsx
@@ -1,5 +1,5 @@
 import { DonutChart } from '@/components/charts'
-import { CompactMetric } from '@/components/ui/compact-metric'
+import { CompactMetric } from '@/components/ui/CompactMetric'
 import { Archive, Code, GitBranch, Star } from 'lucide-react'
 import { fetchAllRepositories } from './github-utils'
