+++ apps/insights/app/github/utils/date-helpers.ts
@@ -1,0 +1,9 @@
+/**
+ * Get the start of the week for a given date (Sunday)
+ */
+export function getWeekStart(date: Date): Date {
+  const d = new Date(date)
+  const day = d.getDay()
+  const diff = d.getDate() - day
+  return new Date(d.setDate(diff))
+}
