+++ apps/insights/app/wakatime/page.tsx
@@ -1,7 +1,7 @@
 import Image from 'next/image'
 import { Suspense } from 'react'
-import { SkeletonCard } from '../../components/skeleton-card'
-import { StaticCard } from '../../components/static-card'
+import { SkeletonCard } from '../../components/SkeletonCard'
+import { StaticCard } from '../../components/StaticCard'
 import { WakaTimeActivity } from './activity'
 import { WakaTimeLanguages } from './languages'
 import { WakaTimeMetrics } from './metrics'
