+++ apps/insights/components/PopularContentTable.tsx
