+++ apps/insights/components/SkeletonCard.tsx
