+++ apps/insights/components/TextDataSource.tsx
