+++ apps/insights/components/charts/AreaChart.tsx
