+++ apps/insights/components/charts/BarChart.tsx
