+++ apps/insights/components/charts/BarList.tsx
