+++ apps/insights/components/charts/LanguageBarChart.tsx
@@ -43,45 +43,9 @@ const chartConfig = {
   },
 } satisfies ChartConfig
 
-// Custom label component for model names
-const ModelNameLabel = (props: LabelProps) => {
-  const { payload, x = 0, y = 0, width = 0, height = 0 } = props
-  
-  // Handle case where payload might be undefined
-  if (!payload) return null
-  
-  // Convert string|number to number
-  const numX = typeof x === 'string' ? parseFloat(x) : x
-  const numY = typeof y === 'string' ? parseFloat(y) : y
-  const numWidth = typeof width === 'string' ? parseFloat(width) : width
-  const numHeight = typeof height === 'string' ? parseFloat(height) : height
-  
-  const isShortBar = payload.isShortBar || (payload.percent || 0) < 15
-  const displayName = payload.displayName || payload.name || ''
-  
-  // For short bars, position text outside (to the right of the bar)
-  // For long bars, position text inside (left side of the bar)
-  const textX = isShortBar ? numX + numWidth + 8 : numX + 8
-  const textAnchor = isShortBar ? 'start' : 'start'
-  const fill = isShortBar ? 'hsl(var(--foreground))' : 'hsl(var(--background))'
-  
-  return (
-    <text
-      x={textX}
-      y={numY + numHeight / 2}
-      textAnchor={textAnchor}
-      dominantBaseline="middle"
-      fontSize={11}
-      fontWeight={500}
-      fill={fill}
-    >
-      {displayName}
-    </text>
-  )
-}
 
-// Custom label component for percentages
-const PercentageLabel = (props: LabelProps) => {
+// Custom label component that combines model name and percentage
+const CombinedLabel = (props: LabelProps) => {
   const { payload, x = 0, y = 0, width = 0, height = 0, value } = props
   
   // Handle case where payload might be undefined
@@ -96,21 +60,26 @@ const PercentageLabel = (props: LabelProps) => {
   
   const isShortBar = payload.isShortBar || (payload.percent || 0) < 15
   const displayName = payload.displayName || payload.name || ''
+  const percentage = `${Number(numValue || payload.percent || 0).toFixed(1)}%`
   
-  // For short bars with external text, position percentage after the model name
-  // For long bars, position percentage at the right edge of the bar
-  const textX = isShortBar ? numX + numWidth + 8 + displayName.length * 6.5 + 10 : numX + numWidth + 8
+  // Combine name and percentage into a single string to avoid positioning issues
+  const combinedText = isShortBar ? `${displayName} (${percentage})` : percentage
+  
+  // For short bars, position outside; for long bars, position at the right edge
+  const textX = isShortBar ? numX + numWidth + 8 : numX + numWidth - 8
+  const textAnchor = isShortBar ? 'start' : 'end'
+  const fill = isShortBar ? 'hsl(var(--foreground))' : 'hsl(var(--background))'
   
   return (
     <text
       x={textX}
       y={numY + numHeight / 2}
-      textAnchor="start"
+      textAnchor={textAnchor}
       dominantBaseline="middle"
       fontSize={12}
-      fill="hsl(var(--foreground))"
+      fill={fill}
     >
-      {`${Number(numValue || payload.percent || 0).toFixed(1)}%`}
+      {combinedText}
     </text>
   )
 }
@@ -152,10 +121,7 @@ export function LanguageBarChart({ data, className }: LanguageBarChartProps) {
         />
         <Bar dataKey="percent" fill="hsl(var(--chart-1))" radius={4}>
           <LabelList
-            content={ModelNameLabel}
-          />
-          <LabelList
-            content={PercentageLabel}
+            content={CombinedLabel}
           />
         </Bar>
       </BarChart>
