+++ apps/insights/components/ui/CompactMetric.tsx
