+++ apps/insights/components/ui/MetricCard.tsx
