+++ apps/insights/components/ui/chart/ChartContainer.tsx
@@ -1,0 +1,44 @@
+'use client'
+
+import * as React from 'react'
+import * as RechartsPrimitive from 'recharts'
+import { cn } from '@/lib/utils'
+import { ChartContext } from './context'
+import type { ChartConfig } from './types'
+
+export const ChartContainer = React.forwardRef<
+  HTMLDivElement,
+  React.ComponentProps<'div'> & {
+    config: ChartConfig
+    children: React.ComponentProps<
+      typeof RechartsPrimitive.ResponsiveContainer
+    >['children']
+  }
+>(({ id, className, children, config, ...props }, ref) => {
+  const uniqueId = React.useId()
+  const chartId = `chart-${id || uniqueId.replace(/:/g, '')}`
+
+  return (
+    <ChartContext.Provider value={{ config }}>
+      <div
+        data-chart={chartId}
+        ref={ref}
+        className={cn(
+          'flex h-[200px] w-full justify-center text-xs',
+          '[&_.recharts-cartesian-axis-tick_text]:fill-muted-foreground',
+          '[&_.recharts-cartesian-grid_line]:stroke-border/20',
+          '[&_.recharts-tooltip-cursor]:stroke-border',
+          '[&_.recharts-surface]:outline-none',
+          className,
+        )}
+        {...props}
+      >
+        <RechartsPrimitive.ResponsiveContainer>
+          {children}
+        </RechartsPrimitive.ResponsiveContainer>
+      </div>
+    </ChartContext.Provider>
+  )
+})
+
+ChartContainer.displayName = 'ChartContainer'
