+++ apps/insights/components/ui/chart/ChartLegend.tsx
@@ -1,0 +1,69 @@
+'use client'
+
+import * as React from 'react'
+import * as RechartsPrimitive from 'recharts'
+import { cn } from '@/lib/utils'
+import { useChart } from './context'
+import { getPayloadConfigFromPayload } from './utils'
+
+export const ChartLegend = RechartsPrimitive.Legend
+
+export const ChartLegendContent = React.forwardRef<
+  HTMLDivElement,
+  React.ComponentProps<'div'> & {
+    payload?: Array<Record<string, unknown>>
+    verticalAlign?: 'top' | 'bottom'
+    hideIcon?: boolean
+    nameKey?: string
+  }
+>(
+  (
+    { className, hideIcon = false, payload, verticalAlign = 'bottom', nameKey },
+    ref,
+  ) => {
+    const { config } = useChart()
+
+    if (!payload?.length) {
+      return null
+    }
+
+    return (
+      <div
+        ref={ref}
+        className={cn(
+          'flex items-center justify-center gap-4',
+          verticalAlign === 'top' && 'pb-3',
+          className,
+        )}
+      >
+        {payload.map((item, index) => {
+          const key = `${nameKey || item.dataKey || 'value'}`
+          const itemConfig = getPayloadConfigFromPayload(config, item, key)
+
+          return (
+            <div
+              key={String(item.value || index)}
+              className={cn(
+                'flex items-center gap-1.5 [&>svg]:h-3 [&>svg]:w-3 [&>svg]:text-muted-foreground',
+              )}
+            >
+              {itemConfig?.icon && !hideIcon ? (
+                <itemConfig.icon />
+              ) : (
+                <div
+                  className="h-2 w-2 shrink-0 rounded-[2px]"
+                  style={{
+                    backgroundColor: String(item.color || '#000'),
+                  }}
+                />
+              )}
+              {itemConfig?.label}
+            </div>
+          )
+        })}
+      </div>
+    )
+  },
+)
+
+ChartLegendContent.displayName = 'ChartLegendContent'
