+++ apps/insights/components/ui/chart/context.ts
@@ -1,0 +1,14 @@
+import * as React from 'react'
+import type { ChartContextProps } from './types'
+
+export const ChartContext = React.createContext<ChartContextProps | null>(null)
+
+export function useChart() {
+  const context = React.useContext(ChartContext)
+
+  if (!context) {
+    throw new Error('useChart must be used within a <ChartContainer />')
+  }
+
+  return context
+}
