+++ apps/insights/components/ui/chart/index.ts
@@ -1,0 +1,7 @@
+// Export all chart components and utilities
+export * from './types'
+export * from './context'
+export * from './utils'
+export { ChartContainer } from './ChartContainer'
+export { ChartTooltip, ChartTooltipContent } from './ChartTooltip'
+export { ChartLegend, ChartLegendContent } from './ChartLegend'
