+++ apps/insights/components/ui/chart/types.ts
@@ -1,0 +1,18 @@
+import * as React from 'react'
+
+// eslint-disable-next-line @typescript-eslint/no-unused-vars
+const THEMES = { light: '', dark: '.dark' } as const
+
+export type ChartConfig = {
+  [k in string]: {
+    label?: React.ReactNode
+    icon?: React.ComponentType
+  } & (
+    | { color?: string; theme?: never }
+    | { color?: never; theme: Record<keyof typeof THEMES, string> }
+  )
+}
+
+export type ChartContextProps = {
+  config: ChartConfig
+}
